requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
pydantic==1.9.0
passlib==1.7.4
pyjwt==2.3.0
psycopg2==2.9.1
python-dotenv==1.0.1
